+++ frontend/beCompliant/src/pages/activityPage/ActivityPage.tsx
@@ -8,20 +8,20 @@ import {
   Text,
 } from '@kvib/react';
 import { useParams } from 'react-router';
-import { Page } from '../components/layout/Page';
-import { TableComponent } from '../components/Table';
-import { TableStatistics } from '../components/table/TableStatistics';
-import { useAnswers } from '../hooks/useAnswers';
-import { useComments } from '../hooks/useComments';
-import { useForm } from '../hooks/useForm';
-import { mapTableDataRecords } from '../utils/mapperUtil';
-import { AnswerType } from '../api/types';
-import { ErrorState } from '../components/ErrorState';
-import { useContext } from '../hooks/useContext';
-import { useUser } from '../hooks/useUser';
+import { Page } from '../../components/layout/Page';
+import { TableComponent } from './table/Table';
+import { TableStatistics } from './TableStatistics';
+import { useAnswers } from '../../hooks/useAnswers';
+import { useComments } from '../../hooks/useComments';
+import { useForm } from '../../hooks/useForm';
+import { mapTableDataRecords } from '../../utils/mapperUtil';
+import { AnswerType } from '../../api/types';
+import { ErrorState } from '../../components/ErrorState';
+import { useContext } from '../../hooks/useContext';
+import { useUser } from '../../hooks/useUser';
 import { useState } from 'react';
-import { SettingsModal } from '../components/table/SettingsModal';
-import RedirectBackButton from '../components/RedirectBackButton';
+import { SettingsModal } from './settingsModal/SettingsModal';
+import RedirectBackButton from '../../components/buttons/RedirectBackButton';
 import { useQueryClient } from '@tanstack/react-query';
 
 export default function ActivityPage() {
