+++ frontend/beCompliant/src/pages/activityPage/TableStatistics.tsx
