+++ frontend/beCompliant/src/pages/activityPage/table/DataTableCell.tsx
