+++ frontend/beCompliant/src/pages/activityPage/table/TableState.tsx
