+++ frontend/beCompliant/src/pages/questionPage/QuestionDetails.tsx
@@ -2,8 +2,8 @@ import { Text, Stack, StackProps, Box, Tag, Flex } from '@kvib/react';
 import { Question } from '../../api/types';
 import Markdown from 'react-markdown';
 import { markdownComponents } from '../../utils/markdownComponents';
-import { LoadingState } from '../LoadingState';
-import { ErrorState } from '../ErrorState';
+import { LoadingState } from '../../components/LoadingState';
+import { ErrorState } from '../../components/ErrorState';
 import { useFetchColumns } from '../../hooks/useFetchColumns';
 import colorUtils from '../../utils/colorUtils';
 import { formatDateTime, isOlderThan } from '../../utils/formatTime';
