+++ frontend/beCompliant/src/pages/questionPage/TextAreaAnswer.tsx
